+++ packages/fighting-design/toolbar-item/src/toolbar-item.vue
@@ -8,25 +8,22 @@
   const prop: ToolbarItemPropsType = defineProps(Props)
 
   /**
-   * 点击触发
-   *
+   * @zh 点击触发
    * @param evt 事件对象
    */
   const handleClick: HandleMouseEventInterface = (evt: MouseEvent): void => {
-    runCallback(prop.click, evt)
+    runCallback(prop.onClick, evt)
   }
 </script>
 
 <template>
   <span
     class="f-toolbar-item"
-    :data-key="dataKey"
+    :data-index="index"
     :style="{ color }"
     @click="handleClick"
   >
-    <f-svg-icon v-if="icon" :size="iconSize">
-      <component :is="icon" />
-    </f-svg-icon>
+    <f-svg-icon v-if="icon" :size="iconSize" :icon="icon" />
     <slot />
   </span>
 </template>